pyproject.toml
[tool.mypy]
mypy_path = "stubs"

warn_unused_ignores = true
warn_return_any = true
show_error_codes = true

strict_optional = true
implicit_optional = true

disallow_any_unimported = true
disallow_subclassing_any = true
#disallow_any_generics = true

disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true

check_untyped_defs = true

[tool.isort]
atomic = true
balanced_wrapping = true
known_first_party = "sr.robot3"
skip_gitignore = true

# hanging grid grouped indent style wrapping
multi_line_output = 5
include_trailing_comma = true

[tool.pytest.ini_options]
markers = [
    "hardware: mark a test that requires hardware to be connected",
]

[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "sr/robot3/_version.py"

[tool.setuptools.packages.find]
include = ["sr", "sr.*", "sr.robot3", "sr.robot3.*"]

[project]
name = "sr-robot3"
description = "Student Robotics API for Python 3"
authors = [{name = "Student Robotics", email =  "kit-team@studentrobotics.org"}]
readme = "README.md"
license = {file = "LICENSE"}
dynamic = ["version"]
requires-python = ">=3.9"
dependencies = [
    "pyserial >=3,<4",
    "april_vision==3.0.0",
    "paho-mqtt >=2,<3",
    "pydantic >=1.9.1,<2",
    "typing-extensions; python_version<'3.10'",
    "tomli >=2.0.1,<3; python_version<'3.11'",
]
classifiers = [
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Typing :: Typed",
    "Topic :: Education",
]

[project.urls]
Repository = "https://github.com/srobo/sr-robot"
Homepage = "https://studentrobotics.org"
Documentation = "https://docs.studentrobotics.org"

[project.optional-dependencies]
dev = [
    "flake8",
    "isort",
    "mypy>=1.7,<2",
    "build",
    "types-pyserial",
    "pytest",
    "pytest-cov",
    "paho-mqtt >=2,<3"
]
vision = ["opencv-python-headless >=4.10,<5"]
